# api/swagger.yaml
swagger: '2.0'
info:
  description: ''
  version: 1.0.0
  title: 挣闲钱
  termsOfService: 'http://swagger.io/terms/'
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: petstore.swagger.io
basePath: /
...
schemes:
  - https
  - http
paths:
  /user/publishQuery:
    post:
      tags:
        - user
      summary: publish questionareTask
      description: publish a new questionareTask(发布一个新的问卷任务，需要填写表单)
      operationId: publishQTask
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Task form information should be posted(任务表单)
          required: true
          schema:
            $ref: '#/definitions/questionareTask'
      responses:
        '201':
          description: create success
          schema:
            $ref: '#/definitions/questionareTask'
        '400':
          description: bad request
          schema:
            type: object
            properties:
              error:
                type: string
...
  /user/accept:
    post:
      tags:
        - user
      summary: accept task
      description: accept a new task(接受一个新任务，需要绑定用户和任务)
      operationId: acceptTask
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          description: the task that user accept(用户和接受的任务)
          required: true
          schema:
            $ref: '#/definitions/userAndTask'
      responses:
        '201':
          description: success
          schema:
            $ref: '#/definitions/userAndTask'
        '400':
          description: bad request
          schema:
            type: object
            properties:
              error:
                type: string
...
  /user/finishAccept:
    delete:
      tags:
        - user
      summary: finish task
      description: finish a accepted task by userId and TaskId(通过用户id和任务id完成一个已接收的任务)
      operationId: finishAccept
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          description: the task that user accept(用户和接受的任务)
          required: true
          schema:
            $ref: '#/definitions/userAndTask'
      responses:
        '204':
          description: success
        '404':
          description: not found
          schema:
            type: object
            properties:
              error:
                type: string
  /user/finishPublish:
    delete:
      tags:
        - user
      summary: finish task
      description: finish a published task(结束/取消一个已发布的任务)
      operationId: finishPublish
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          description: the task
          required: true
          schema:
            $ref: '#/definitions/task'
      responses:
        '204':
          description: success
        '404':
          description: not found
          schema:
            type: object
            properties:
              error:
                type: string
...
  '/user/{userId}':
    get:
      tags:
        - user
      summary: get user by userId
      description: ''
      operationId: getUser
      produces:
        - application/json
      parameters:
        - name: userId
          in: path
          description: ''
          required: true
          type: integer
      responses:
        '200':
          description: success
          schema:
            $ref: '#/definitions/user'
        '404':
          description: not found
          schema:
            type: object
            properties:
              error:
                type: string
                example: Wrong UserId or Password
definitions:
  task:
    type: object
    properties:
      taskId:
        type: integer
      taskType:
        type: string
      taskTitle:
        type: string
      endTime:
        type: string
      userId:
        type: integer
  questionare:
    type: object
    properties:
      queryid:
        type: integer
      content:
        type: string
...
  user:
    type: object
    properties:
      userId:
        type: integer
      username:
        type: string
      password:
        type: string
      university:
        type: string
      grade:
        type: string
      phone:
        type: string
      major:
        type: string
      email:
        type: string
      balance:
        type: integer
...
  questionareTask:
    properties:
      task:
        $ref: '#/definitions/task'
      questionare:
        $ref: '#/definitions/questionare'
...
  userAndTask:
    properties:
      taskId:
        type: integer
      userId:
        type: integer
...